Added time-masking to the SimpleRNN model.
This introduces a new mode to the SimpleRNN where it can be
"time-masked" in which case the input becomes doubled along the time
dimension, with X[:,X.shape[1]//2:,0] becoming a mask vector which, when
set to 0, compels the _scan function to simply copy its h_tm1 to the
output.

Also, Make tests into a module.

This allows me to run the tests via, e.g.
`python -m tests.test_masked_recurrent`
which in turn allows me to test against a development version of Keras
that I've not installed.