layers.recurrent: add masking of all-zeroes timesteps
If there is no input layer to a recurrent layer, the layer can
automatically induce the masking from the data if `mask_zero` is True.

all-zeroes timesteps are assumed to be padding, and are masked out
in training or prediction. If `mask_zero` is False and there is no
previous layer with output_mask, then all input is assumed to be
non-padding.

Solves issue #355.

Signed-off-by: Amit Beka <amit.beka@gmail.com>