Early stopping on growing log
Hello, 
I just made this pr to show a idea.  The early stopping was made to be applied on loss or decreasing values. 
I'm just proposing to letting it be applied on growing values like accuracy or val_acc. That being said  I don't know if this is a bad pattern or not.

BTW : congratulations guys ! Keras is awesome you do an excellent job !