Added support for user modification of the number of threads that the tensorflow backend uses to run its "intra_op_parallelism". This was done by making it so that the tensorflow backend respects the setting of the environment variable OMP_NUM_THREADS, which is usually used to set the number of threads that a BLAS package runs with. If the environment variable is not set, then tensorflow runs with its default behavior, which is to use the maxmum number of threads that are accessible.
The reasons to limit the number of threads are twofold. First, one
might wish to run multiple instances of keras concurrently on the
same machine, which would be impeeded by a process that consumes all
of the avilable computational resources. Second, there are cases where
the software actually runs faster on fewer cores, such as when the
overhead for locking and communicating between parallel processes
scales faster than the computuational benefit of the parallel
processes. This was observed on a machine with two 8-core
(16 thread) Intel(R) Xeon(R) CPU E5-2660 0 @ 2.20GHz CPUs. Using more
than 8 threads at a time resulted in a significant slowdown using the
lstm_text_generation example. Using fewer threads in BLAS alone did
not improve performance, and neither did decreasing the number of
threads in tensorflow while leaving BLAS alone. It is because of
this relationship that the number of threads available to
tensorflow is set to track the number of threads available to BLAS.
Additionally, this makes the behavior of the tensorflow backend
equivalent to the behavior of the theano backend, which already
adjusts its performance implicly based on the setting of the
OMP_NUM_THREADS environment variable.

In short, this improves the user control, performance, and
consistency of the keras package.