add weight sharing across layers
Weight sharing is currently only implemented by Siamese, which can be problematic when one shared layer depends on the output of another.
This PR hijacks the `weights` parameter supported by the constructor of most existing layers.
If a sharedvariable is passed instead of a numpy variable, it overwrites the weight member itself (not its value!).