[Merge] use layer_cache
Currently the Merge layer calls directly `layer.get_output()`.
This can lead to higher compilation times as the `layer_cache` is
not used. For a model with many `Merge` layers this can be quite an
issue. With this commit the `Merge` layer checks first,  if the output is
already in the `layer_cache`.