Fixes #2110 "ValueError: I/O operation on closed file"
This is a workaround for #2110 where calling `model.fit`
with
`verbose=1` using IPython can intermittently raise "ValueError: I/O
operation on closed file".

This exception appears to be caused by an unknown IO bug with IPython
that is raised when updating the `ProgbarLogger` progress bar . To
workaround this bug and prevent users from unexpectedly losing their
model:

- The minimum progress bar refresh interval is now 0.1 seconds (up from
0.01 seconds).

- Progress bar updates are now wrapped in `try/catch` blocks that ignore
`ValueError` exceptions raised when calling `progbar.update`

An ideal solution would resolve the IPython at the source, however, this
is an important workaround for users who want to use IPython with
`verbose=1`.