+imdb_multilevel_gru example
Multi-level RNNs are getting more popular, important especially (but not
only) for character level models. Moreover, they are a bit unintuitive
in Keras when they are bidirectional. This example demonstrates how to
construct them, a few common tricks, and a few exercises that might
prove educational for newcomers to RNNs for NLP.