Add LSTM with timegate
Adding a timegate to the LSTM architecture showed great promise at this year's NIPS conference. 
A simple example on MNIST dataset is shown here, compared to standard LSTM in Keras: https://github.com/fferroni/PhasedLSTM-Keras/blob/master/README.md 

I am wondering if it could become part of the Keras library, and if so, what are the requirements for the code to be accepted.

Thank you for the great library.