Update optimizers.py
This edit allows initializing optimizers so as to start training from a specific iteration. This will be helpful if training is interrupted for some reason and the user wants to continue from where the training was interrupted.
Without this edit, even a cold reloading from a saved weight give some slight difference in error/accuracy values.