Checks if model.stop_training is true after on_batch_end in fit_generator
Without this change, setting model.stop_training = True in on_batch_end of a callback has no effect when using the fit_generator. This behaviour is different when using fit. This change makes fit and fit_generator respond alike to a training stop at the end of a batch in a callback.