Prevent loading HD5Stores into memory
Keras provides HD5Stores to store large dataset on disc and treat it as a numpy array. But the command np.asarray loads the whole thing into memory causing an out of memory error. This change fixes that, as long as your HD5Store is in float32 or whatever dtype keras is set to.

(why not just use .flow_from_directory?, keras uses PIL which only support images with three channels)