Increase file save hash size
Issue #8249 runs into a problem where the max hash size is low, resulting in files being potentially overwritten.

The behavior of this functionality is as such:
As batch size increases, the probability that any given file will have a conflict in the hash reduces. However, when the author wants to save images with a batch size of one, the probability of a conflict increases since {index} is always 1. This seems to just be an unforeseen bug because a batch size of one is uncommon.