fix list_devices() function is not available issue (#8567)
* fix list_devices() function is not available issue

When using tensorflow as backend and the version of tensorflow is under r1.3, the list_devices() function is not available of Session instance. This may cause some issue like keras.utils.multi_gpu_model(model, gpus) function can not work under r1.3 of tensorflow. This change is a hack to fix that issue.

* Update tensorflow_backend.py

fix pep8 check failure

* Update tensorflow_backend.py

follow fchollet review comment. use `if not`.