Use tensorflow leaky_relu op for efficiency (#9044)
* Use tensorflow leaky_relu op for efficiency

The current implementation for leaky_relu is extremely inefficient. In my specific usecase it took as much time as the convolution itself and led to tensorslow being much slower than theano. The old inefficient implementation was the result of tensorflow not having the leaky_relu op, but it was recently added.

* fix max_value clipping

* fix pep8