Add tests to batch_dot
I added three tests to test the backend.batch_dot() function. They all yield inconsistent behavior on Theano and Tensorflow backends. I have not tested on CNTK backend because I don't have it running on my machine.

The first test throws error on Theano but not on Tensorflow because Theano does not allow summing over the 0th index in batch_dot while Tensorflow does. The second test throws indexing error on Tensorflow (see this issue: https://github.com/keras-team/keras/issues/9847) and the same error as the first test on Theano. However, there might be some tensor shapes for which the second test works on Theano but not on Tensorflow, but I couldn't think of one immediately. The third test yields tensors of different shapes on the two backends.

I am not sure what is the intended behaviour of batch_dot() in these cases I have illustrated so I couldn't add tests to test_batch_dot_shape(). Perhaps the documentation should be updated to indicate the desired behaviour more clearly so that the functionality could be compared to that.