adds uniform_zoom to ImageDataGenerator
This modification gives the user the ability to define, if the zoom transformation should be the same on the xy axes or seperate for each axis. This are additional changes for github.com/keras-team/keras-preprocessing/pull/44.

Hope this is the right way to suggest such change. Sorry if not, but I never did this before :)