Call set_model on callbacks during load_model
The load_model function builds the model from the saved object, and
implicitly sets the model in the backed when the weights are set.

Callbacks that require set_model to be called before the model is
set on the backend can fail to operate on loaded models as a result.

This commit adds the ability to pass callbacks to the load_model function.