### Summary
Information from the docstrings of the file constraints.py was not grabbed by autogen.py because the autogenerated code block was absent in templates/constraints.md and the corresponsing part of the PAGES list in autogen.py was not there too. I did not find a commit where it would work before, but i could have missed it.

### Related Issues
Resolves the issue #11382
### PR Overview

- [n] This PR requires new unit tests [y/n] (make sure tests are included)
- [y] This PR requires to update the documentation [y/n] (make sure the docs are up-to-date)
- [y] This PR is backwards compatible [y/n]
- [n] This PR changes the current API [y/n] (all API changes need to be approved by fchollet)