use multiprocessing Event instead of manager Event
There is a race condition during failure/shutdown when using a stop
event instance from the multiprocessing manager class.  The manager can
exit before the main process has a chance to check the stop event state,
but with the manager dead this will result in EOF/ConnRefused. Using the
stop event from the multiprocessing module makes it independent and
avoids the race condition