Delete misleading comment about Dense input flattening
From this comment it sounds like there is an implicit
`Flatten` layer before the dot product with `kernel`.

However, this seems misleading to me, as the output
shape of this layer is the same as the input shape
with the last dimension replaced by `units`. This
is also mentioned further down in the comment, so
I propose just deleting this comment about the
flattening.

Alternatively, we could replace it by a comment that
says something about the dot product only being done
along the last dimension.