Running on Python 3.6 on OSX:
If a model file doesn't exist, I was having this error thrown:

Unable to create group (no write intent on file)

I then had an error with:

'Group' object has no attribute 'decode'

According to https://stackoverflow.com/questions/28583565/str-object-has-no-attribute-decode-python-3-error

This error was generated due to decoding a string that is already in UTF-8 and is safe to ignore.

This patch fixes both of these things.