Added intersection over union metric for bbox
Bounding Box should be in the format x1,x2,y1,y2. x1,y1 left upper corner of the box in the image and x2,y2 be the right  lower corner of the box in the image. This metric is very useful as it can be used in many object localization tasks.