Remove argument `unroll` from `ConvRNN2D`
Argument `unroll` is not supported in `ConvRNN2D` as it 1s written in the docs that it's not possible with Convolutional RNNS. But the argument itself still exists and causes error in TRAVIS test_doc: **ValueError: ("ConvRNN2D ['unroll'] arguments are not present in documentation ", 'keras.layers.convolutional_recurrent')**. 

There are two ways to fix this error: 1. add `unroll` method explanations in `ConvRNN2D`document, 2. remove `unroll` argument from `ConvRNN2D` code. I chose the latter, because an argument which has no other selectives doesn't need to be noted for user convinience and simplicity(as it was in docs before this fix), and `RNN`, the parent `ConvRNN2D`, owns `unroll=False` as its default, so `unroll` of ConvRNN2D stays fixed as `False`. Currenty `unroll` arguement does nothing but putting out a TypeError when `unroll=True`, which can't happen if there's no `unroll`.