Clarify axis in normalization.py
The phrase "axis that should be normalized" is ambiguous and unclear.

See: https://stackoverflow.com/a/47549590/6296435

Many people understand it to be taking mean and std **along** that axis, i.e. collapse that dimension and preserve all other dimensions. This is also the behavior of np.mean, and other operations.

But this is obviously the opposite of what is going on here, and so the documentation should be corrected to reflect this.