Update searchbox.html
When visiting a site that returns a 404 error, for example "keras.io/atreatthth", using the searchbar functionality results in the browser navigating to "https://search.html/?q=SEARCH_TEXT". This was documented in DOCS BUG #13736.

This is due to the searchbar form "action" attribute being set to "//search.html" on the 404-returning sites. 

By changing the searchbox to always be "action="./search.html"", instead of {{base_url}} /search.html, I believe this issue can be avoided.