Add new target for tensorflow_core which doesn't contain keras.
Historically tensorflow_no_contrib contains keras since its treated as a core API, and estimator/tb relies on it as a build dependency. When we want to move keras out as a component, we need a target that only contains tf core code (doesn't include keras), and keras could rely on it as a build dependency.

To keep the status quo, I add the both tf.keras and oss keras to tensorflow_no_contrib, so that tf.keras API will be still available at this target.

PiperOrigin-RevId: 356829561