Skip histograms in Keras automatic outside compilation test if not using MLIR bridge
A pending change to the histogram summary (https://github.com/tensorflow/tensorboard/pull/5093) breaks the Keras automatic outside compilation test when run without the MLIR bridge. To work around this, we disable the histogram-specific parts of the test in that case. The longer term fix for this is https://github.com/tensorflow/tensorboard/issues/2885.

This also disables the legacy copy of this test within the TensorFlow repo entirely (rather than backporting the fix and other changes it depends on).

PiperOrigin-RevId: 385040526