First draft of functional model sliding.
The interface for building functional model from tensor that is not tf.keras.Input is same as the existing tf.keras.Model(). User just need to feed in any existing keras_tensor as inputs to the model when constructing it.

Under the hood, the functional model will check if any of the inputs are intermediate keras_tensor, and create clone for related inputs/output/nodes for new model.

PiperOrigin-RevId: 386960334