Fix metric correctness issue in distributed setting.
It is unclear to my why the SyncOnRead variable behaves differently for __add__ and tf.math.add(). v1 + v2 will return pre-replica value and tf.math.add(v1, v2) will return aggregated value. This is causing related metrics to calculate wrong result when metric is created in the MirroredStrategy scope. So far the issue is only noticed for mirrored strategy and works fine for TPUStrategy.

PiperOrigin-RevId: 388820436