This changes the default for `tpu_py_test` from `disable_mlir_bridge = True` to `disable_mlir_bridge = False`.
Before this change, calling `tpu_py_test` with `disable_mlir_bridge = True` was actually using the MLIR bridge, since rollout of MLIR bridge changed the bridge default.

Note that when the default is passed to `tpu_py_test`, only one bridge is tested. When the non-default is passed, both bridges are tested. This cl preserves this property.

PiperOrigin-RevId: 390253471