Dynamically add Trackable methods to the LossScaleOptimizer.
As SavedModel refactors are going on, there are many updates to the methods in Trackable. Since Keras lives in a different repo from core, it's hard to update this file at the same time as the refactor is happening. This CL makes it so that this file dynamically reflects the updates.

PiperOrigin-RevId: 391661835