Improve discoverability of preprocessing layers among keras layers
With the new symbol location of preprocessing layers directly in the keras layers
namespace, we need to be more explicit in the doc strings that these layers
are preprocessing layers.

We also link the preprocessing layer guide, which is the only convinient way
to find the full set of these layers on tensorflow.org.

PiperOrigin-RevId: 391873446