Don't rely on the per-batch times to compute the per-epoch `steps_per_second`.
When using `tf.keras.Model.fit` and `ParameterServerStrategy`, `on_batch_end` is not called when the batch finishes, but rather when it has been enqueued.

This change adds `_epoch_start_time`, which is set to the current time in `on_epoch_begin`, and used to compute the total epoch time in `on_epoch_end`.

PiperOrigin-RevId: 393332359