Add idf_weights as an init argument for lookup layers
When lookup layers are in tf_idf output mode, a set of inverse document frequency
weights can be learned via adapt, or set via set_vocabulary as an idf_weights
argument. This adds idf_weights as an init arg to the lookup layers, which allows
specifying a vocabulary in tf_idf output mode on layer construction.

There was no reason not to be doing this already, this was just a gap.

PiperOrigin-RevId: 396875253