Information from masked attention inputs can leak
Hi,

I discovered while working on an MNIST attention demo, that a network is able to learn adequate reconstruction performance on MNIST test set, even when the entire input is masked out.

When masking, before the softmax operation in the MultiHeadAttention layer the masked inputs are set to -1E9. This results in masked value vectors having a negligible contribution to the final sum.

However, the network is able to learn to distinguish between the resulting *very small* floating point numbers of the masked out values, which appears as inexplicable good performance.

The fix (for self attention) is to multiply the attention outputs by the mask. This PR just puts a warning in the docs, but I encourage someone to think of a better solution.