Allow controlling output dtype of categorical preprocessing layers
When output_mode is "one_hot", "multi_hot", or "count" output dtype will follow
the global policy unless a dtype is explicitly set on the layer.

When output_mode is "int" output dtype will default to int64, unless a dtype is
explicitly set on the layer. An explicit floating point output will lead to an
error.

PiperOrigin-RevId: 400835665