Improve SavedModel error message when loading a layer fails
We should not consume all ValueErrors and assume they fit the case where a
custom layer has not been registered. A ValueError on creating a layer could
reflect a number of issues and should be surfaced directly.

Before this change, all value errors:
RuntimeError: Unable to restore a layer of class TextVectorization. Layers of class
TextVectorization require that the class be provided to the model loading code,
either by registering the class using `@keras.utils.register_keras_serializable`
on the class def and including that file in your program, or by passing the class
in a `keras.utils.CustomObjectScope` that wraps this load call.

After this change, unkown layer:
ValueError: Unknown layer: SomeLayer. Please ensure this object is passed to the
`custom_objects` argument. See
https://www.tensorflow.org/guide/keras/save_and_serialize#registering_the_custom_object
for details.

After this change, unkown callable arg:
ValueError: Unkown value for `standardize` argument of layer TextVectorization.
If restoring a model and `standardize` is a custom callable, please ensure the
callable is registered as a custom object. See
https://www.tensorflow.org/guide/keras/save_and_serialize#registering_the_custom_object
for details. Allowed values are: `None`, a `Callable`, or one of the following
values: lower_and_strip_punctuation. Received: some_function
PiperOrigin-RevId: 401873404