Subclass StringLookup directly from TextVectorization
This adds the same dtype support for TextVectorization that was recently added to
StringLookup. It also avoids duplication of logic around saving, vocab setting/loading,
and saved configs. This should leave the layer less fragile to bugs where it is not
updating in sync with StringLookup.

This also removed an undocumented output_mode=None setting that appears to have been
repected on the layer only for testing.

PiperOrigin-RevId: 402353962