Add a initial_best parameter in ModelCheckpoint
Add a initial_best parameter in ModelCheckpoint that is stored as self.best. if a session is stopped and rerun or kernel dies, then the best value of metric is taken as np.inf or -np.inf and the model is saved after the first epoch without taking into consideration the initial performance of model.