Keras Metric.
Implement class `OneHotMeanIoU` that inherits class `MeanIoU`.

This class can be used to compute the mean IoU metric for a multi-class classification task where both the labels are one-hot encoded.

This class will first convert both the labels and predictions into the original format of the labels, then applies the computation steps of class `MeanIoU`.

PiperOrigin-RevId: 409065377