Fix RandomCrop layer called on integer input
We had a tf.cond which called tf.image.resize on one path (returning float32) and
cropping the input dtype on the other, which lead to a error. Instead we cast the
inputs to the layer's compute dtype and do our computation with that type.

Fixes https://github.com/keras-team/keras/issues/15681

PiperOrigin-RevId: 417686461