Update losses.py
There are some inconsistencies in the format of documentation. In some places, "`y_true` and `y_pred`" words are used, whereas "labels and predictions" are used in others. Although these two terms mean the same thing, the discrepancy disrupts the document's general format and may cause confusion for newcomers. So I think it should be changed to bring one specific order.