Update BaseImageAugmentationLayer for transformation that are input dependent.
For KPL that need to coordinate between image/label/bbox, it might need to generate transformation based on the input (eg shape of the image, number of bbox, etc). The current method doesn't provide any information to user.

In future, when we need to add more input types (eg segmentation masks), we will add a new optional param to the `get_random_transformation` method

PiperOrigin-RevId: 437098775