Validate file integrity on `get_file` on download.
Previously, the file integrity was checked only if the target path existed, as a mean to prevent redowloading the same file.

However, security conscious users would get a hash of the file from a separate channel (e.g., from the page hosting the model if it also displays the checksum) and would use that to make sure that the model they download is exactly the one they intend to. This is essential to prevent malicious downloads / corrupted ones / MitM type of attacks.

PiperOrigin-RevId: 437329179