Add arg `custom_shape` to method `add_variable_from_reference()` so that users can control the shape of optimizer variables.
This is related to b/226547716: in optimizer like compact adagrad, the optimizer variable is by intention set as a different shape from the model variable to reduce memory cost: go/alise-ml/compact_adagrad. Thus, we need a way to let users control the shape of added variable, but meanwhile we also want optimizer variables to be placed at the same device as model variable.

PiperOrigin-RevId: 438437215