Add a new mode for backend.RandomGenerator.
The new stateless mode is intended to be used by keras.Initializer, so that the generator will not contain or refer to any tf.Variable. This will cause a behavior change to seedless initializer, which will now produce same value if it is called multiple times (with same shape). The initializer change will come in a follow up cl.

PiperOrigin-RevId: 444370618