Address the issue of "partition_offset" being ignored by keras initializers.
With the upcoming stateless initializer change, the current behavior of
keras initializer will cause issue for unseeded initializer in PSS,
which call the initializer multiple times with potentially same shape
and different offset. We need to pass the offset information to keras
backend RNG and let it produce different result for different offset
value. Since the offset is only provided at call(), we need to add
extra param to all the methods in keras RNG.

PiperOrigin-RevId: 444929801