Introducing pre-commit to the repo
Pre-commit is installed as part of the project dependencies
in the development container and pre-commit environment is initialized
after container is successfully created.

Hooks are set to check for general style violations,
along with forgotten debug statement and merge resolution leftovers.
The checks will be run after relevant files are next edited,
to keep git-blame output clean.

Signed-off-by: Jiri Podivin <jpodivin@gmail.com>