Remove caching of constants as member variables in BatchNormalization.
In graph mode, tensors are specific to the graph they were created in, and using a tensor from a different graph is not allowed. When "fused" mode is enabled, a constant 1 and 0 tensor of an appropriate shape is cached as a member variable in BatchNormalization. This can trigger cross-graph bugs when using a BatchNormalization instances from different graph functions.

To fix this, this change stores the param shape as a member variable and creates a ones or zeros tensor of that shape on the fly.

Impact: This should be a low-impact change as it only affects the case where `center` or `scale` are False when fused batchnorm is in use, and they are both on by default.
PiperOrigin-RevId: 456562374