Modified `TimeDistributed` to not require `compute_output_shape` to be implemented by the wrapped layer, and to not use `set_shape` on the output.
The output shape can be inferred from the output of the wrapped layer combined with the known batch size and number of time steps. The same applies to the mask.
The know static batch size is used in the final reshape instead of -1 so that the static shape of the output has the batch dimension defined, which removes the need for `set_shape`.

https://github.com/keras-team/keras/issues/16656

PiperOrigin-RevId: 465692038