DTensorCheckpointV2 [6/n] Keras Checkpoint Restoration for Late Variable Creation
This CL starts a suite of tests for Keras integration with DTensorCheckpointV2. DTensorCheckpointV2 essentially is an effort to not use DTensorCheckpoint api anymore, and instead fully integrate DTensor with tf.train.Checkpoint so that there is dtensor special api for checkpointing.

This CL includes a test for late variable creation restoration through add_weight api.

PiperOrigin-RevId: 466738889