Added a optional argument
Added an optional "Factor" argument (set to 1 by default). The outputs will be multiplied by this factor. This is to regulate and prevent numbers increasing or decreasing exponentially when multiple dense layers are in sequence. I have personally run into this issue. I originally solved it using lambda layers, but the underlying functions became hard to keep track of at around the fifth function. This should solve that problem for larger networks.