Adding DTensor support for non-sync batchnorm.
This is a quite complicated case, since the mean/var need to be used:
1. Update the moving mean/var which is a state tracking variable (need to reduce the
 global tensor shape).
2. Normalize the output and we need to keep the batch dimension so that each of the element will get the correct (unreduced mean/var).

The renorm during training is not support since it is too complicated and a very cornered use case. I choose to raise an explicit error message for the moment.

The unit test case has been update to compare the result between the existing strategy and the new strategy to make sure we are mathematically correct.

PiperOrigin-RevId: 514565995