Add support for keras optimizer to run under dtensor based strategy.scope.
When running with DTensor based strategy, users don't need to explicitly provide mesh, and the variables will  be properly created as DVariable (because of the scope), and the other aggregration/reduce/update methods should go through the dtensor logic as well.

The dtensor training test has been update to test optimizer creation under the strategy scope.

PiperOrigin-RevId: 518612615