Updating split range from [0, 1] to (0, 1) in tf.keras.utils.split_dataset to make it more intuitive
In the split_dataset API which is used to split a dataset into left and right parts. But documentation states it splits dataset into left half and right half where the parts need not be equal parts.The split can be anything in range(0, 1).Hence I think to replace the word half with part.
Also in the argument section and exception section the range for split used as [0, 1] which is not a correct notation as 0 and 1 are actually excluded in the range and raises exception if we use either 0 or 1 for either of left_size or right_size for splitting the dataset. Hence i have replaced [0, 1] with (0,1) to make it more intuitive. Thanks.