State `y_pred` must be 2 dimensional for AUC.
For a multilabel AUC metric, we expect that for each data point we will have one or multiple labels. This results in a two dimensional array. Currently, when `y_pred` is three dimensional we incorrectly state that `y_true` must have two dimensions. The resulting shape will be the same whether it is taken from `y_pred` or `y_true`, but `y_true` is not always available at compile time.

PiperOrigin-RevId: 528485014