Use `tf.where` to replace masked-out values with negative infinity.
Currently, we subtract a large value (32752 for float16, 1e9) from masked-out inputs to set them to negative infinity. However, this does not work if the input is positive infinity or NaN, in which case the input remains as-is. This results in NaNs being propogated through to the output, as the softmax of an input containing infinities or NaNs is NaN for all output values. This PR explicitly replaces masked-out values with negative infinity and adds tests to verify this behavior.

PiperOrigin-RevId: 539687392