Always save the build config when building a layer (#820)
* Always save the build config when building a layer

This is trying to put us in a more consistent place with respect to
implicitly vs explicitly built layers. Explicitly building will now...
- Record the build config passed to the layer.
- Lock the layer so state cannot be mutated.

The last remaining different between implicitly vs explicitly build is
that implicitly we will attempt the build the layer by calling it
symbolically, explicitly we do not. We could revisit that if we wanted,
but I think that would be a slightly more complex change.

* More simplifications