Add `named_product` test utility to generate test cases with names.
This is the be used with `absl.parameterized.named_parameters` and instead of `absl.parameterized.product`.
It creates testcases that have intuitive names.

Tests in `numpy_test.py` and `dtypes_test.py` that were using `product` now use `named_product`.

Also standardized the way we create `KerasTensors` in `numpy_test.py` for readability.