Add `tf.SparseTensor` and `tf.IndexedSlices` support to a number of TensorFlow ops.
The following element-wise unary ops now support `tf.SparseTensor` and `tf.IndexedSlices`. The output is of the same type as the input.
- abs
- absolute
- arcsin
- arcsinh
- arctan
- arctanh
- ceil
- conj
- conjugate
- copy
- expm1
- floor
- imag
- log1p
- negative
- real
- round
- sign
- sin
- sinh
- sqrt
- square
- tan
- tanh

The following element-wise unary ops now support `tf.SparseTensor` and `tf.IndexedSlices`. The output is dense.
- arccos
- arccosh
- cos
- cosh
- exp
- log
- log10
- log2
- reciprocal

The following element-wise binary ops now support `tf.SparseTensor` and `tf.IndexedSlices`. The output type depends on the two inputs and the op.
- add
- subtract
- maximum
- minimum
- multiply
- mod
- divide
- true_divide
- floor_divide

The following reduction op now supports `tf.IndexedSlices`. The output is an `tf.IndexedSlices` unless dimension 0 is reduced or the rank of the output is 1 or less.
- mean

This is in preparation for supporting sparse gradients in optimizers.