Restore hacky workaround for tf dict wrapper errors
In https://github.com/keras-team/keras/pull/18507 we removed a hacky
workaround for functional models so dictionary output is not made a
trackable.

We actually still need this when calling `predict` on a functional with
dictionary output.

Added testing to plug the coverage gap. Still no idea what the status
of the underlying bug is :/, filed with tensorflow

https://github.com/tensorflow/tensorflow/issues/62217