Fix for layers.Reshape when using -1 and a dynamic batch size. (#18792)
When using `-1` for one of the dimensions and a dynamic batch size, with the Tensorflow backend, in non eager mode, this would result in Tensors with a `None` dimension where the `-1` was. This is because the `reshape` op does not treat the batch size specially and therefore has no way to determine what goes to the batch dimension vs. the `-1` dimension.

To fix this, we resolve the `-1` to an actual number at build time since the layer knows that the batch dimension is untouched.

Note that this bug would not affect the JAX and PyTorch backends because we inject a fake batch size when needed.