Change implementation of numpy.min() of torch backend (#19066)
* Change implementation of numpy.min() of torch backend

The implementation of min() with torch backend is not working properly. For eg if list of axis passed the reduction is not happening. Changed the implementation like torch.numpy.max function which is working fine.

Attached https://colab.research.google.com/gist/SuryanarayanaY/9b28fd5fa5837d11d660550caa13eea0/19064.ipynb for demo of same.

* Added test case for  numpy.min function

Added test case for  numpy.min function to test when axis is list or axis is empty.This is to verify the same behaviour of torch wrt other backends also.

* Rectify lint errors