Attempt to add support for saving/loading bfloat16
Currently attempting to save or load weights in bfloat16 will fail.
There may be better ways to do this, but the approach jax and tf seem
to take is to use the ml-dtypes library to allow bfloat16 to work with
numpy.

This is further compounded by the h5py format, which saves bfloat16 as
a void type. This implementation currently just assumes any two byte
void type is bfloat16, which seems a bit hacky. Quite possibly a better
way to do this.