Unwrap variable values in all stateless calls.
When variables are passed to stateless calls (`stateless_call`, `stateless_update`, `stateless_apply`), unwrap them to extract their value so that the mapping from variable to value in `StatelessScope` points to a value.
This is to prevent an infinite recursion when performing operations (e.g. `__add__`) in a stateless scope.

Also fix issue where casting a `tf.SparseTensor` would lose the shape. The optimization in `ops.cast` is what revealed the stateless calls bug.