Fix bug in `model.export()` when inputs in a structure.
The code computing the `call` signature was not going through the structure correctly.
`TrackedDict` and `_DictWrapper` do not play nice with `jax2tf`/JAX so we need to turn them to plain Python collections when using `_build_shapes_dict`.

Also tweaked tracking in `Layer` class as the first attribute set was not tracked.