Don't mutate passed config dict in Operation.from_config()
As of #19728, Operation.from_config() has been overriding the dtype
field in the passed config dict.
There doesn't seem to be an explicit contract that from_config doesn't
modify the config argument, but this seems to be the case in practice,
and this behaviour has broken some code on an internal repo.