Implement tool for saved Keras model files inspection, diff and patching
It consists of three sub features, that allow:
Visualization of the contents of .keras and .weights.h5 files, where you can further expand the contents up until the weights of the file.
Diff of a model compared to a reference model, being presented side by side with the differences highlighted.
Patching a model, where it is possible to change a layer's name and changing a specified weight of a model.

Co-authored-by: Pedro Curto <pedro.a.curto@tecnico.ulisboa.pt>