Loading using legacy h5 api now calls load_own_variables
https://github.com/keras-team/keras/issues/20147

```
So, I investigated it a bit and the problem is the keras.layers.Normalization layer, which is not finalized.
This is because keras.layers.Normalization has two weights that are loaded from file adapted_mean and adapted_std, but this are not used directly when making a prediction.
What is used directly are two variables, which are not added as weights, called mean and std, that are updated whenever finalize_state() is called.

And this is the problem. Theoretically, whenever loading weights from file, finalize_state() should be called by load_own_variables().
But load_own_variables() is called only with the new saving API, when the weights are saved in the format ".weights.h5", while the efficientnet weights have the legacy extension ".h5".

Therefore, weights are loading, we have adapted_mean and adapted_std, but the actual mean and std are the one initialized with build(), i.e. a vector of zeros and a vector of ones.

As a workaround, simply calling base_model.layers[2].finalize_state() will fix the issue.```