Allow `ops.stop_gradient()` to take a variable. (#20302)
Passing a tensor to `ops.stop_gradient()` always works. However, passing a variable directly would work with the Tensorflow backend but fail with an obscure error message with the JAX backend and the Torch backend, requiring users to write `ops.stop_gradient(variable.value)`.

This makes it work directly with variables, which is a common use-case.