Fix `set_tensor_attr` bug related to garbage collection.
WeakKeyDictionary is really what we wanted for this, but the issue is that tensors are not hashable and can't be used as keys.

By making values weakly referenced, we can run in two issues:
- After a mask is set, typically nothing keeps a reference to the mask, so if garbage collection happens, all masks are likely lost.
- After a tensor used as key is garbage collected, nothing notifies WeakValueDictionary because the id is just an id an not connected to the tensor object, so the key/value pair is not removed from the dictionary. When the id is reused, a random tensor can get paired with a random preexisting mask.

This fix uses finalizers to be notified that a tensor is garbage collected to clear its mask.