Major rework of `optree_impl` and `dmtree_impl` for consistency.
The `optree` implementation and the `dmtree` implementation of the `tree` API had a nummber of discreptancies. Running unit tests without `optree` installed would fail on a number of tests.

The documentation and behavior of the `tree` API was not internally consistent. There was contradicting documentation about the handling of `OrderedDict`s. The behavior of the `optree` implementation was to use the key sorted order for `pack_sequence_as` but use the sequence order for `flatten`, resulting in `flatten` + `pack_sequence_as` not being idempotent (as discovered in https://github.com/keras-team/keras/issues/20538 )

The exceptions used to report non-matching structures where different between the two implementation. When `optree` uses `ValueError` for all mimatches, `dmtree` would distinguish between `ValueError` and `TypeError` for some cases. This caused a number of bugs because `TypeError` was often not caught, only `ValueError`.

The `assert_same_structure` argument of `check_types` is deprecated and no longer does anything. The `optree` implementation would check the types of the *leaves*, whereas the `dmtree` would check the types of the *collections*. So `check_types=False` with `optree` was fairly, although not completely similar, to `check_types=True` with `dmtree`. The rules is that no two collection types are considered the same, except for `dict`, `OrderedDict` and `defaultdict`.

Because `optree` is the default implementation used and `dmtree` is only a fallback, this PR changes the `tree` API behavior to conform to the `optree` approach everywhere. This makes the `optree` implementation a thin wrapper on top of `optree`, whereas large portions of the `dmtree` wrapper are now reimplemented in Python. Note that the `tree` API was initially modelled after `dmtree`, not `optree`.

There are a couple of fairly niche known discrepancies between the `optree` implementation and the `dmtree` implementation. They are documented in `dmtree_impl.py`.

- Fixed references to `unflatten_as` in documentation, which doesn't exist.
- Fixed contradicting documentation in `flatten` and `pack_sequence_as` related to the handling of `OrderedDict`s. The documentation now states that the sequence order is always used.
- Made handling of `OrderedDict`s follow the spec with both `optree` and `dmtree`.
- Made the exceptions raised consistent and documented them. `TypeError` is only for major programmer error (e.g. `func` is not callable), and `ValueError` is used for all structure mismatches.
- Removed all cases where `TypeError` was caught after a `assert_same_structure`.
- Fixed the discrepancy in the behavior for `namedtuple`s. The `optree` behavior is now followed, meaning that the path for fields are indices, not field names.
- Deprecated the `check_types` argument in `assert_same_structure` and implemented the `optree` behavior in `dmtree`.
- Remove `sequence_fn` argument of `pack_sequence_as`, which was not used and force the `optree` implementation to be fully rewritten in Python.
- Added `MAP_TO_NONE` to the API, added support for it in both implementations of `traverse`. This feature was documented, but not accessible and not actually implemented.
- Added support for registered classes with `dmtree`, both with `flatten` and `unflatten` passed at registration time and methods on the class.
- Tracked collections are now supported with `dmtree` (`TrackedList`, `TrackedSet` and `TrackedDict`). In particular, `TrackedSet` would be handled as a leaf and never traversed.
- Removed dependency of tracked collections on `optree` in `tree_flatten` and `tree_unflatten`.
- Tensorflow's `ListWrapper` and `_DictWrapper` are now supported with `dmtree`.
- Implemented a more efficient way for `optree` to verify structures are the same while traversing them with `map_structure` and `map_structure_up_to`. This avoids multiple traversals.
- Added documentation for `list_to_tuples` and `map_shape_structure`.
- Completely rewrote all tree API unit tests, which are now painfully thorough.
- `map_shape_structure` is now unit tested.
- Fixed unintented use of `tree` instead of `keras.tree` in unit test.
- Ran unit tests for all backends with `optree` uninstalled.

Fixes https://github.com/keras-team/keras/issues/20538