Add support for symbolic tensors to `convert_to_tensor`. (#20539)
`convert_to_tensor(x, sparse=False)` is the API to densify sparse tensors. When used in that manner, the input is already a backend tensor. For this scenario, it makes sense to support symbolic tensors so that one can build a functional model using `convert_to_tensor`.

Also improved the documentation of `convert_to_tensor`.